Flower patch menu

See merge request cis4820/Beeverse!4

+++ Assets/Scripts/GameState/BeeResources.cs
@@ -1,7 +1,7 @@
 public struct BeeResources {
-  public int honey;
-  public int beeswax;
-  public int royalJelly;
-  public int nectar;
-  public int pollen;
+  public float honey;
+  public float beeswax;
+  public float royalJelly;
+  public float nectar;
+  public float pollen;
 }
